+++ NEWS.md
@@ -3,6 +3,7 @@
 - [x] difference views
 - [ ] restatement/idempotency unit tests
 - [x] all types of 3NF views in dashboard
+- [x] sequence class to reflects the SQL sequences
 - [ ] cube - batch join based on model
 - [ ] load/view supports shared knots
 - [ ] load/view supports 2+ role ties